+++ views/index.hbs
@@ -114,3 +114,5 @@ body{
 </script>
 
 <script> require('../routes/app.js');</script>
+<script> require('../routes/login.js');</script>
+
